+++ .gitignore
@@ -1,17 +1,9 @@
 /data
 /database_east/
-/unsupervised_line_segmentation/my_way/cutted_pages/
-/unsupervised_line_segmentation/my_way/normalized/
-/unsupervised_line_segmentation/my_way/whole_pages/
 /unsupervised_line_segmentation/dataset/
 /unsupervised_line_segmentation/my_way/s_histograms/
 /zzz_map_work/
-/unsupervised_line_segmentation/my_way/patches/
-/unsupervised_line_segmentation/my_way/cutted_pages_backup/
-/unsupervised_line_segmentation/my_way/grayscale_dataset_prepared_200px/
-/unsupervised_line_segmentation/my_way/grayscale_dataset_prepared_128px/
-/unsupervised_line_segmentation/my_way/grayscale_dataset/
-/unsupervised_line_segmentation/my_way/grayscale_dataset_prepared_150px/
-/unsupervised_line_segmentation/my_way/grayscale_dataset_150px_s098/
-/unsupervised_line_segmentation/my_way/grayscale_dataset_150px_s050/
 /line-seg_matlab/
+/unsupervised_line_segmentation/my_way/data/
+/unsupervised_line_segmentation/my_way/models/
+/line_seg_matlab/
