--- preprocessing.py
@@ -1,4 +1,0 @@
-import numpy as np
-import tensorflow as tf
-
-print(tf.config.list_physical_devices('GPU'))
