--- unsupervised_line_segmentation/my_way/learned_model/log
@@ -1,16 +1,0 @@
-epoch,accuracy,loss,val_accuracy,val_loss
-0,0.5,5.658689427375793,0.49000000953674316,0.9420642185211182
-1,0.46,0.7589727544784546,0.5099999904632568,0.6956585597991943
-2,0.51,0.6989051866531372,0.800000011920929,0.6840596771240235
-3,0.54,0.6898988080024719,0.49000000953674316,0.677726902961731
-4,0.57,0.6527316236495971,0.49000000953674316,0.7883762395381928
-5,0.69,0.5859668803215027,0.49000000953674316,0.525266021490097
-6,0.59,0.592009916305542,0.7599999904632568,0.5684660220146179
-7,0.68,0.5751349806785584,0.7900000214576721,0.46341741204261777
-8,0.75,0.5075662970542908,0.49000000953674316,0.5665425705909729
-9,0.72,0.5062990963459015,0.7200000286102295,0.5212903833389282
-10,0.76,0.4937616753578186,0.7900000214576721,0.47615497827529907
-11,0.79,0.4514084482192993,0.8299999833106995,0.4294544252753258
-12,0.79,0.45878496408462527,0.7799999713897705,0.4388011100888252
-13,0.79,0.4304918700456619,0.8100000023841858,0.43047007262706755
-14,0.79,0.4212019371986389,0.8199999928474426,0.4112608468532562
